+++ internal/handler/web/users.go
@@ -1,0 +1,69 @@
+package handler
+
+import (
+	"net/http"
+	"slices"
+
+	"github.com/gin-gonic/gin"
+	"github.com/rs/zerolog/log"
+	"github.com/ucl-arc-tre/portal/internal/middleware"
+	openapi "github.com/ucl-arc-tre/portal/internal/openapi/web"
+	"github.com/ucl-arc-tre/portal/internal/types"
+
+	"github.com/ucl-arc-tre/portal/internal/rbac"
+)
+
+// GetPeople retrieves all people in the system, but only if the user has admin role.
+func (h *Handler) GetPeople(ctx *gin.Context) {
+	user := middleware.GetUser(ctx)
+	roles, err := rbac.GetRoles(user)
+	if err != nil {
+		setServerError(ctx, err, "Failed to get roles for user")
+		return
+	}
+
+	if slices.Contains(roles, "admin") {
+		// retrieve auth + agreements + training info
+		people, err := h.users.GetAllPeople()
+		if err != nil {
+			setServerError(ctx, err, "Failed to get people")
+			return
+		}
+		ctx.JSON(http.StatusOK, people)
+
+	} else {
+		ctx.JSON(http.StatusForbidden, gin.H{"error": "Forbidden"})
+	}
+
+}
+
+func (h *Handler) GetProfile(ctx *gin.Context) {
+	user := middleware.GetUser(ctx)
+	attributes, err := h.users.Attributes(user)
+	if err != nil {
+		log.Err(err).Msg("Failed to get user attributes")
+		ctx.Status(http.StatusInternalServerError)
+		return
+	}
+	ctx.JSON(http.StatusOK, openapi.ProfileResponse{
+		Username:   string(user.Username),
+		ChosenName: string(attributes.ChosenName),
+	})
+}
+
+// Updates the user's chosen name.
+func (h *Handler) PostProfile(ctx *gin.Context) {
+	user := middleware.GetUser(ctx)
+	update := openapi.ProfileUpdate{}
+	if err := ctx.ShouldBindJSON(&update); err != nil {
+		setInvalid(ctx, err, "Invalid JSON object")
+		return
+	}
+	if err := h.users.SetUserChosenName(user, types.ChosenName(update.ChosenName)); err != nil {
+		setServerError(ctx, err, "Failed to update chosen name")
+		return
+	}
+	ctx.JSON(http.StatusOK, openapi.ProfileUpdate{
+		ChosenName: update.ChosenName,
+	})
+}
